+++ engine/app/components/citizens_advice_components/footer.rb
@@ -41,7 +41,6 @@ module CitizensAdviceComponents
       FooterFeedbackLink.new(
         url: @feedback_url,
         title: t("citizens_advice_components.footer.website_feedback"),
-        external: true,
         new_tab: true
       )
     end
@@ -65,39 +64,24 @@ module CitizensAdviceComponents
     end
 
     class FooterColumn < Base
-      attr_reader :title, :links, :icon
+      attr_reader :title, :links
 
       def initialize(title:, links:)
         super
         @title = title
         @links = links
-
-        icon_option_deprecation
-      end
-
-      def icon_option_deprecation
-        return unless links.any? { |link| link[:icon].present? }
-
-        CitizensAdviceComponents.deprecator.warn(
-          "generic `icon` property for column links is deprecated use `external: true` instead"
-        )
       end
     end
 
     class FooterFeedbackLink < Base
       attr_reader :title, :url
 
-      def initialize(url:, title: nil, external: false, new_tab: false)
+      def initialize(url:, title: nil, new_tab: false)
         super
 
         @url = url.to_s
         @title = title
-        @external = external
         @new_tab = new_tab
-      end
-
-      def external?
-        @external.present?
       end
 
       def new_tab?